Cameo/Aphid: Fix overly picky PRU1 state machine timing
Waiting for a particular coincidence of \PCMD high and PR/\W low could lock up PRU0 for
nearly 11 seconds if non-handshake activity on the control lines wiggled \PCMD and PR/\W
together in a confusing way (specifically, in a way that left PR/\W high or had PR/\W go
from low to high very quickly). The initialisation routine for the parallel interface
card was prone to generating these toxic wiggles, but owing to vagaries of signal
propagation and other electrical mysteries, a Cameo/Aphid device would only be affected
by it if it was plugged into the top port. This change relaxes the coincidence check so
that a \PCMD-raise with PR\/W high terminates the handshake immediately---the state
machine no longer waits for a while for PR\/W to go low again.